1) Исправлен баг когда юзер мог сам добавить себе роль 2) Исправлен баг с огромным UserUpdateRequest 3) Из всех DTO убрана лишняя информация (например в тестах юзера информация о создателе) 4) Добавлено взаимодействие с авторизацией

+++ src/main/java/ru/vibelab/tplatfom/services/UserService.java
@@ -137,6 +137,10 @@ public class UserService {
     }
 
     public void deleteUser(UserDeleteRequest request) {
+        Long id = request.getId();
+        if (id == 1) {
+            throw new UserNotFoundException("Unable to delete Admin");
+        }
         userRepo.delete(userRepo.findById(request.getId()).orElseThrow(
                 () -> new UserNotFoundException(request.getId())));
     }
